清除编译告警

+++ utils/codec/src/asn1_node.cpp
@@ -333,7 +333,7 @@ int32_t Asn1Node::Asn1AsBits()
     }
     uint32_t bits = static_cast<uint32_t>(dataBits);
     int32_t index = static_cast<int32_t>(dataLength_) - 1;
-    for (index; index < sizeof(int32_t); index++) {
+    for (index; index < static_cast<int32_t>(sizeof(int32_t)); index++) {
         bits <<= OFFSET_EIGHT_BIT;
     }
     integerVal = static_cast<int32_t>(Asn1Utils::ReverseInt(bits));
